apps/api/requirements.txt
# KillKrill API Dependencies
# Quart-based backend with PyDAL

# Web Framework
quart>=0.19.0
quart-cors>=0.7.0
hypercorn[h3]>=0.16.0

# Database
pydal>=20241031.1
psycopg2-binary>=2.9.9

# Redis
redis>=5.2.1
hiredis>=2.3.2

# Authentication
pyjwt>=2.10.1
passlib>=1.7.4
bcrypt>=4.2.1

# Validation
pydantic>=2.7.0
email-validator>=2.1.0

# HTTP Client
httpx[http3]>=0.28.1
aiohttp>=3.9.0

# Monitoring
prometheus-client>=0.21.1
structlog>=24.5.0

# Configuration
python-decouple>=3.8

# Utilities
python-dateutil>=2.9.0
pytz>=2024.2
uuid6>=2024.1.12

# License Integration
requests>=2.32.4

# WebSocket
websockets>=12.0
